Feat: Contextapi + useReducer로 vm상태관리(state-mangement)
commit 8da9a64b76017a97e41e137a91eec4a7ea0e9e37
Author: yunho <kimyouknow@naver.com>
Date:   Wed May 11 16:38:25 2022 +0900

    Bug: GNB 버그 주석 추가

    url과 상관없이 활성화탭이 기본 값으로 됨.

commit 99310e0f6431c41ef11ca67cd93e143a464613bd
Author: yunho <kimyouknow@naver.com>
Date:   Wed May 11 16:34:00 2022 +0900

    Feat: VMmangement페이지에 Products 컴포넌트 추가

commit 378c7ffa1c933f38da570b66ee0ec99fc11e721e
Author: yunho <kimyouknow@naver.com>
Date:   Wed May 11 16:33:38 2022 +0900

    Refactor: VendingMachine 페이지와 VMContext연결

commit a701919177bf90116aee4457213f450724f37ae6
Author: yunho <kimyouknow@naver.com>
Date:   Wed May 11 16:33:05 2022 +0900

    Feat: Products 컴포넌트 분리

commit 1538afb7cd8115c8645df529b13f4f00f762313a
Author: yunho <kimyouknow@naver.com>
Date:   Wed May 11 16:18:47 2022 +0900

    Feat: VMContext

commit ac89f2a1d418e602b8ce00bf02666be13de9e6e5
Author: yunho <kimyouknow@naver.com>
Date:   Wed May 11 16:06:45 2022 +0900

    Refactor: 재고관리 잔돈 누를시 줄어드는 로직 제거
