+++ fe/src/pages/VMmangement/ChangesUnits.jsx
@@ -1,30 +1,16 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { Button } from '@mui/material';
 import { setKoreanLocaleMoney } from 'utils';
 import * as S from './style';
 
-function ChangesUnits({ totalBalance, changesUnits, substactBalance }) {
-  const handleClickUnit = useCallback(
-    unit => {
-      const targetUnit = changesUnits.find(unitElement => unitElement.unit === unit);
-      if (targetUnit.count <= 0) {
-        alert('잔고가 없어요');
-        return;
-      }
-      substactBalance(unit);
-    },
-    [changesUnits, substactBalance],
-  );
-
+function ChangesUnits({ totalBalance, changesUnits }) {
   return (
     <S.ChangesUnitsContainer>
       <div>{setKoreanLocaleMoney(totalBalance)}원</div>
       {changesUnits.map(({ id, unit, count }) => (
         <li key={id}>
-          <Button variant="contained" onClick={() => handleClickUnit(unit)}>
-            {unit}
-          </Button>
+          <Button variant="contained">{unit}</Button>
           <div>{count}</div>
         </li>
       ))}
@@ -41,7 +27,6 @@ ChangesUnits.propTypes = {
       count: PropTypes.number.isRequired,
     }),
   ).isRequired,
-  substactBalance: PropTypes.func.isRequired,
 };
 
 export default ChangesUnits;
