+++ fe/src/pages/VMmangement/style.js
@@ -1,5 +1,10 @@
-/* eslint-disable import/prefer-default-export */
 import styled from 'styled-components';
+
+export const Container = styled.div`
+  display: flex;
+  width: 100%;
+  gap: 36px;
+`;
 
 export const ChangesUnitsContainer = styled.ul`
   width: 20%;
